add test for lint pass, move rename tests
